campaing send

+++ app/Services/CampaignService.php
@@ -13,7 +13,7 @@ class CampaignService extends Base
         parent::__construct();
     }
 
-    public function create_campaign(array $data)
+    public function create_campaign(array $data): array
     {
         $contactListIds = $data["contact_list_ids"];
 
@@ -31,14 +31,14 @@ class CampaignService extends Base
         return [true, MCH_model_created("Campaign"), [$campaign], 200];
     }
 
-    public function show_all_campaigns()
+    public function show_all_campaigns(): array
     {
-        $campaigns = $this->user->campaigns()->with('contactLists')->orderBy("created_at", "desc")->get();
+        $campaigns = $this->user->campaigns()->with('contactLists')->orderBy("created_at", "desc")->get()->toArray();
 
         return [true, MCH_model_retrieved("Campaigns"), $campaigns, 200];
     }
 
-    public function update_campaign(Campaign $campaign, array $data)
+    public function update_campaign(Campaign $campaign, array $data): array
     {
         if (isset($data["contact_list_ids"])) {
             $contactListIds = $data["contact_list_ids"];
@@ -47,7 +47,7 @@ class CampaignService extends Base
             unset($data["contact_list_ids"]);
         }
 
-        if (isset($data['status']) && $data['status'] === CampaignStatusEnum::PUBLISHED) {
+        if (isset($data['status']) && $data['status'] === CampaignStatusEnum::PUBLISHED->value) {
             if ($campaign->template === null) {
                 return [false, CAMPAIGN_TEMPLATE_NOT_SET, [], 400];
             }
@@ -55,15 +55,19 @@ class CampaignService extends Base
 
         $campaign->update($data);
 
-        $campaign->refresh();
+        $campaign->fresh();
 
         $campaign->publish_campaign();
 
         return [true, MCH_model_updated("Campaign"), [$campaign], 200];
     }
 
-    public function delete_campaign(Campaign $campaign)
+    public function delete_campaign(Campaign $campaign): array
     {
+        if ($campaign->status !== CampaignStatusEnum::DRAFT) {
+            return [false, CAMPAIGN_TEMPLATE_NOT_SET, [], 400];
+        }
+
         $campaign->delete();
 
         return [true, MCH_model_deleted("Campaign"), [], 200];
